# packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: Telephone-Booth Operator API
  description: |
    Versioned REST + WebSocket API between the operator console
    (`Telephone-Booth-Operator`) and the phone-side Rust client.

    The phone client authenticates with a static Bearer API token issued from
    the operator UI. Operator browser sessions authenticate via Authentik OIDC
    (cookie session). Native mobile apps in
    `Telephone-Booth-Operator-Mobile` authenticate with Authentik PKCE and send
    the resulting access token as `Authorization: Bearer <jwt>`
    (`operatorBearer` security scheme).
  version: 1.0.0
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8787
    description: Local dev
  - url: https://operator.example.com
    description: Production
...
paths:
...
  /v1/status:
    get:
      summary: Current booth status as last reported by the phone client.
      tags: [status]
      security:
        - operatorSession: []
        - operatorBearer: []
        - apiToken: []
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoothStatus" }
    put:
      summary: Phone client posts a status update.
      tags: [status]
      security:
        - apiToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StatusUpdate" }
      responses:
        "204": { description: accepted }
...
  /v1/stats/summary:
    get:
      summary: Aggregated booth status + queue counters for mobile widgets.
      description: |
        Cached aggregation suitable for high-frequency widget polling.
        Returns the latest booth status, message-queue counts (pending,
        received-today, latest id), call counts (today + in-progress), and
        the active WebSocket subscriber count. Cached for 5 seconds on the
        server so 1 request per 15s widget timeline ≈ 0 extra DB load.
      tags: [stats]
      security:
        - operatorSession: []
        - operatorBearer: []
      parameters:
        - name: installationId
          in: query
          description: |
            Installation scope. Omit for the currently active installation
            (the default), pass a uuid for a historical one, or `all` to span
            every installation.
          schema:
            oneOf:
              - { type: string, format: uuid }
              - { type: string, enum: [all] }
      responses:
        "200":
          description: aggregated booth statistics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatsSummary" }
        "401": { description: missing or invalid credentials }
  /v1/stats/overview:
    get:
      summary: Historical usage aggregation over a time window or custom range.
      description: |
        Thorough aggregates across calls, messages, booth events and uploads.
        Select either a preset `window`, or a custom range via `start`/`end`.
        `end` accepts the literal `now` (or omit it) to mean the current
        instant, so a saved custom filter stays live. Preset windows are
        cached; custom ranges are computed fresh.
      tags: [stats]
      security:
        - operatorSession: []
        - operatorBearer: []
      parameters:
        - name: installationId
          in: query
          description: |
            Installation scope. Omit for the currently active installation
            (the default), pass a uuid for a historical one, or `all` to span
            every installation.
          schema:
            oneOf:
              - { type: string, format: uuid }
              - { type: string, enum: [all] }
        - name: window
          in: query
          schema: { $ref: "#/components/schemas/StatsWindow" }
        - name: start
          in: query
          schema: { type: string, format: date-time }
        - name: end
          in: query
          description: An ISO timestamp, or the literal "now".
          schema:
            anyOf:
              - { type: string, format: date-time }
              - { type: string, enum: [now] }
      responses:
        "200":
          description: aggregated usage statistics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatsOverview" }
        "400": { description: invalid window or range }
        "401": { description: missing or invalid credentials }
  /v1/stats/filters:
    get:
      summary: List the current operator's saved metric filters.
      tags: [stats]
      security:
        - operatorSession: []
        - operatorBearer: []
      responses:
        "200":
          description: saved filters
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/MetricFilter" }
        "401": { description: missing or invalid credentials }
    post:
      summary: Save a new metric filter for the current operator.
      tags: [stats]
      security:
        - operatorSession: []
        - operatorBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MetricFilterInput" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MetricFilter" }
        "400": { description: invalid selection }
        "401": { description: missing or invalid credentials }
  /v1/stats/filters/{id}:
    get:
      summary: Fetch one of the current operator's saved filters.
      tags: [stats]
      security:
        - operatorSession: []
        - operatorBearer: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: the saved filter
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MetricFilter" }
        "401": { description: missing or invalid credentials }
        "404": { description: not found or not owned by the caller }
    put:
      summary: Replace one of the current operator's saved filters.
      tags: [stats]
      security:
        - operatorSession: []
        - operatorBearer: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MetricFilterInput" }
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MetricFilter" }
        "400": { description: invalid selection }
        "401": { description: missing or invalid credentials }
        "404": { description: not found or not owned by the caller }
    delete:
      summary: Delete one of the current operator's saved filters.
      tags: [stats]
      security:
        - operatorSession: []
        - operatorBearer: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: deleted }
        "401": { description: missing or invalid credentials }
        "404": { description: not found or not owned by the caller }
...
components:
  securitySchemes:
    apiToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Phone-client API token issued from the operator UI.
    operatorSession:
      type: apiKey
      in: cookie
      name: __Host-booth_session
      description: HMAC-signed opaque session cookie set after OIDC login.
    operatorBearer:
      type: http
      scheme: bearer
      bearerFormat: jwt
      description: |
        Authentik-issued JWT access token. Validated against the provider's
        JWKS, issuer, audience (operator web client_id + any configured
        `OIDC_MOBILE_AUDIENCES`), and operator group / email allow-lists.
        Used by native mobile apps that authenticate via PKCE.
  schemas:
...
    BoothState:
      type: string
      enum:
        [
          idle,
          dialTone,
          dialing,
          playingQuestion,
          beep,
          recording,
          uploading,
          playingMessage,
          playingInstructions,
          callUnavailable,
          error,
        ]
    BoothStatus:
      type: object
      required: [state, updatedAt]
      properties:
        id:
          type: integer
          description: >-
            Snapshot row id. Two runs of the same status can share a booth
            timestamp, so clients use this to tell rows apart; it increases with
            insertion order.
        state: { $ref: "#/components/schemas/BoothState" }
        updatedAt: { type: string, format: date-time }
        firstSeenAt:
          type: string
          format: date-time
          description: >-
            First time the booth reported this status. Identical consecutive
            reports (heartbeats) are collapsed into one snapshot spanning
            firstSeenAt..updatedAt.
        repeatCount:
          type: integer
          minimum: 1
          description: Number of identical booth reports collapsed into this snapshot.
        currentQuestionId: { anyOf: [{ type: string, format: uuid }, { type: "null" }] }
        currentMessageId: { anyOf: [{ type: string, format: uuid }, { type: "null" }] }
        lastError: { anyOf: [{ type: string }, { type: "null" }] }
        runtimeMode:
          anyOf:
            - { type: string, enum: [real, mock, simulator] }
            - { type: "null" }
          description: >-
            How the booth is being driven. Part of the status identity, so a
            change here starts a new snapshot rather than collapsing into the
            current one.
    StatusUpdate:
      type: object
      required: [state]
      properties:
        state: { $ref: "#/components/schemas/BoothState" }
        updatedAt: { type: string, format: date-time }
        currentQuestionId: { anyOf: [{ type: string, format: uuid }, { type: "null" }] }
        currentMessageId: { anyOf: [{ type: string, format: uuid }, { type: "null" }] }
        lastError: { anyOf: [{ type: string }, { type: "null" }] }
        runtimeMode:
          anyOf:
            - { type: string, enum: [real, mock, simulator] }
            - { type: "null" }
          description: >-
            How the booth is being driven. Part of the status identity, so a
            change here starts a new snapshot rather than collapsing into the
            current one.
...
    StatsSummary:
      type: object
      required: [booth, messages, calls, realtime, generatedAt]
      properties:
        booth: { $ref: "#/components/schemas/BoothStatus" }
        messages:
          type: object
          required: [pending, awaitingModeration, receivedToday, latestId]
          properties:
            pending: { type: integer, minimum: 0 }
            awaitingModeration:
              type: integer
              minimum: 0
              description: >
                Messages awaiting operator moderation (status "received" or
                "pending"). Drives the mobile app-icon / tab badge.
            receivedToday: { type: integer, minimum: 0 }
            latestId:
              anyOf:
                - { type: string, format: uuid }
                - { type: "null" }
        calls:
          type: object
          required: [today, inProgress]
          properties:
            today: { type: integer, minimum: 0 }
            inProgress: { type: integer, minimum: 0 }
        realtime:
          type: object
          required: [wsClients]
          properties:
            wsClients: { type: integer, minimum: 0 }
        generatedAt: { type: string, format: date-time }
...
    MetricFilter:
      type: object
      required: [id, name, window, start, end, createdAt, updatedAt]
      description: >
        A per-operator saved metrics time selection. Either `window` (a preset)
        is set, or a custom range via `start`/`end`. A null `end` means "now".
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 80 }
        window:
          anyOf:
            - { $ref: "#/components/schemas/StatsWindow" }
            - { type: "null" }
        start: { anyOf: [{ type: string, format: date-time }, { type: "null" }] }
        end: { anyOf: [{ type: string, format: date-time }, { type: "null" }] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    StatsWindow:
      type: string
      enum: [24h, 7d, 30d, all]
...